stock_backend/requirements.txt
# Django 핵심 패키지
Django>=4.2.0
djangorestframework>=3.14.0

# Django Channels (WebSocket 지원)
channels>=4.0.0
daphne>=4.0.0

# 데이터베이스
psycopg2-binary>=2.9.0

# CORS 지원
django-cors-headers>=4.0.0

# 데이터 처리
pandas>=2.0.0

# 외부 API
finance-datareader>=0.9.50
requests>=2.31.0
websocket-client>=1.6.0

# 환경변수 관리
python-dotenv>=1.0.0

# 프로덕션 서버
gunicorn>=21.0.0 